Further Additions to the Knight
This commit now adds  the ability for the knight to take pieces from the other team, while making sure that it can only be done in the "L" movement that knights are allowed to move in.

The commit also adds new bug fixes to the code.

Although, with these new bug fixes, the code has become very hard to navigate, as well as becoming very inefficient. Therefore, the following commit will have to focus on decreasing the amount of code in the file, and making it easier to navigate.