Restrictions to Movement
The new code additions make sure that the player is not able to move pieces across other friendly pieces.

Meaning, if they want to move a rook 4 squares ahead, they were previously able to do that because the move was considered valid as long as it was straight ahead, and the space was either empty of filled with an enemy piece.

What the new code, its checked if there are friendly pieces in front of the chosen piece. If there are, then the move is disallowed.