Alternations to Functions
The changes to the code here were that the switch function, and what is to occur if the conditions of the "team_kill" function aren't met, and if they're met.

Furthermore, the first implamentation of the function in the code has been made, which included some alterations to the original code.